Fix resource cleanup in server rescue tests
If server rescue tests are skipped then the resource cleanup method
terminates with AttributeError: type object has no attribute 'volume'.
The fix checks whether volume exist before operating with it.

+++ tempest/api/compute/servers/test_server_rescue_negative.py
@@ -56,7 +56,8 @@ class ServerRescueNegativeTestJSON(base.BaseV2ComputeTest):
 
     @classmethod
     def resource_cleanup(cls):
-        cls.delete_volume(cls.volume['id'])
+        if getattr(cls, 'volume', None):
+            cls.delete_volume(cls.volume['id'])
         super(ServerRescueNegativeTestJSON, cls).resource_cleanup()
 
     def _detach(self, server_id, volume_id):
